requirements.txt
google-genai>=1.0.0
Pillow>=10.0.0
reportlab>=4.0
